\name{bs.boot.chisq}
\alias{bs.boot.chisq}
\alias{bsboot}
\alias{summary.bsboot}
\alias{plot.bsboot}
\title{Bollen-Stine Bootstrap}
\description{ Methods to generate and handle the Bollen-Stine Bootstrap. }

\usage{bs.boot.chisq(model, model.data.frame, R=2000, maxiter=500, steptol=1e
-06, t=10, ...)

\method{summary}{bsboot}(object)

\method{plot}{bsboot}(object, chisq.scale=0.1,
		main="", xlab="", ylab="frequency",...)
}

\arguments{
  \item{model}{A ram model object generated by \code{\link{specify.model}}}
  \item{model.data.frame}{A data frame containing only the observed variables in the model.}
  \item{R}{Number of boostrap replicated.  Bollen and Stine Recommend 2000 to get an accurate estimate.}
  \item{maxiter}{The maximum number of iterations for \code{\link{sem}}.  Defaults to 500}
  \item{steptol}{The step tolerance for the \code{\link{nlm}}}
  \item{t}{The tolerance for number of repeats of a single bootstrap to make sure the model was fit.  Defaults to 10}
  \item{...}{Additional arguments for \code{\link{sem}}}
  \item{chisq.scale}{The step size in a \code{link{seq}} statment for calculating a chi squared distribution}
}

\details{This function calculates the bootstrapped chi square value using the method of Bollen and Stine.  Briefly, rather than generate bootstrap replicates of the raw data, the data is transformed, such that the transformed data, D, is defined as

\deqn{D = z * inv(chol(S)) * chol(S)}

where z is the raw data with each variable centered around its mean.  S is the covariance matrix of the raw data.  And \code{link{chol}} means compute the choleski factorization of a matrix.

\code{summary} returns the mean and variance around the Chi Squared value, as well as its mean p value.  \code{plot} plots the density of the bootstrapped chi square value in red against the chi squared distribution with the same degrees of freedom.

Note, all information criteria (aic, aicc, caic, and bic) are also available for \code{bsboot} objects.
}


\seealso{
	\code{\link{sem}}.
}

\references{
Bollen, K.A. and Stine, R.A.  1992.  Bootstrapping goodness of fit measures in structural equation models.  \emph{Sociological Methods Research}.  21: 205-229.



}