\name{combine.models}
\alias{combine.models}
\title{Combine different RAM models}
\description{ I needed to write more code to make working with sem easier. }

\usage{combine.models(...)}
\arguments{
  \item{...}{A series of models in the form of ram objects generated by \code{\link{specify.model}}}
}

\details{Takes multiple models as defined by \code{\link{specify.model}} and smooshes them together into one model.  Can be very useful for working on multiple models which share some, but not all, large structural elements.}


\seealso{
	\code{\link{sem}}
	\code{\link{specify.model}}
}


\examples{ 

    ## Same model as in \link{specify.model}.
    ## broken into two pieces

model.dhp.paths <- specify.model()
    RParAsp  -> RGenAsp, gam11,  NA
    RIQ      -> RGenAsp, gam12,  NA
    RSES     -> RGenAsp, gam13,  NA
    FSES     -> RGenAsp, gam14,  NA
    RSES     -> FGenAsp, gam23,  NA
    FSES     -> FGenAsp, gam24,  NA
    FIQ      -> FGenAsp, gam25,  NA
    FParAsp  -> FGenAsp, gam26,  NA
    FGenAsp  -> RGenAsp, beta12, NA
    RGenAsp  -> FGenAsp, beta21, NA
	RGenAsp  -> ROccAsp,  NA,     1
    RGenAsp  -> REdAsp,  lam21,  NA
    FGenAsp  -> FOccAsp,  NA,     1
    FGenAsp  -> FEdAsp,  lam42,  NA

model.dhp.error<-add.to.model(model.dhp.paths)
    RGenAsp <-> RGenAsp, ps11,   NA
    FGenAsp <-> FGenAsp, ps22,   NA
    RGenAsp <-> FGenAsp, ps12,   NA
    ROccAsp <-> ROccAsp, theta1, NA
    REdAsp  <-> REdAsp,  theta2, NA
    FOccAsp <-> FOccAsp, theta3, NA
    FEdAsp  <-> FEdAsp,  theta4, NA


model.dhp<-combine.models(model.dhp.paths, model.dhp.error)

model.dhp
    ## End(Not run)
    
   }