\name{delete.model.element}
\alias{delete.model.element}
\title{Delete Parts of a Model}
\description{ Deletes a single path in a model }

\usage{delete.model.element(delete.text, old.model, sort.by="path")}
\arguments{
  \item{delete.text}{A character string with either a path, the name of a variable, or the name of a coefficient you wish to remove from a model}
  \item{old.model}{A model in the form of a ram object generated by \code{\link{specify.model}}}
  \item{sort.by}{What kind of information are you giving the delete function?  Is this a path (e.g. a -> b), a whole variable from your model, or the name of a path coefficient?  Options are "path", "variable", or "coefficient" respectively.}
}

\details{Removes selected elements from a model and returns a new ram object.  Only works on one link at a time, unless you are deleting a variable from a model or multiple paths share the same coefficient.}


\seealso{
	\code{\link{sem}}
	\code{\link{specify.model}}
}


\examples{ 

# -------------------- Wheaton et al. alienation data ----------------------
    

S.wh <- read.moments(names=c('Anomia67','Powerless67','Anomia71',
                                    'Powerless71','Education','SEI'))
   11.834                                    
    6.947    9.364                            
    6.819    5.091   12.532                    
    4.783    5.028    7.495    9.986            
   -3.839   -3.889   -3.841   -3.625   9.610     
  -21.899  -18.831  -21.748  -18.775  35.522  450.288

# This is the model in the SAS manual for PROC CALIS: A Recursive SEM with
# latent endogenous and exogenous variables.
# Curiously, both factor loadings for two of the latent variables are fixed.

model.wh.1 <- specify.model()
    Alienation67   ->  Anomia67,      NA,     1
    Alienation67   ->  Powerless67,   NA,     0.833
    Alienation71   ->  Anomia71,      NA,     1
    Alienation71   ->  Powerless71,   NA,     0.833 
    SES            ->  Education,     NA,     1     
    SES            ->  SEI,           lamb,   NA
    SES            ->  Alienation67,  gam1,   NA
    Alienation67   ->  Alienation71,  beta,   NA
    SES            ->  Alienation71,  gam2,   NA
    Anomia67       <-> Anomia67,      the1,   NA
    Anomia71       <-> Anomia71,      the1,   NA
    Powerless67    <-> Powerless67,   the2,   NA
    Powerless71    <-> Powerless71,   the2,   NA
    Education      <-> Education,     the3,   NA
    SEI            <-> SEI,           the4,   NA
    Anomia67       <-> Anomia71,      the5,   NA
    Powerless67    <-> Powerless71,   the5,   NA
    Alienation67   <-> Alienation67,  psi1,   NA
    Alienation71   <-> Alienation71,  psi2,   NA
    SES            <-> SES,           phi,    NA
    
                        
sem.wh.1 <- sem(model.wh.1, S.wh, 932)


#now make wh2 without re-entering the whole thing
model.wh.2<-delete.model.element("Alienation71   ->  Powerless71", model.wh.1, sort.by="path")

model.wh.2<-delete.model.element("Alienation67   ->  Powerless67", model.wh.2, sort.by="path")

model.wh.2 <- add.to.model(model.wh.2)
    Alienation67   ->  Powerless67,   lamby,    NA
    Alienation71   ->  Powerless71,   lamby,    NA 


sem.wh.2 <- sem(model.wh.2, S.wh, 932)

summary(sem.wh.2)


# Compare the two models by a likelihood-ratio test:

anova(sem.wh.1, sem.wh.2)

##    LR Test for Difference Between Models
##    
##            Model Df Model Chisq Df LR Chisq Pr(>Chisq)
##    Model 1        9     13.4851                       
##    Model 2        8     12.6731  1   0.8119     0.3676
    
   }
