\name{sbchisq}
\alias{sbchisq}
\title{Satorra-Bentler Chi Square}
\description{ Calculates an the Satorra-Bentler adjusted Chi Square value }

\usage{sbchisq(sem.obj, sem.data, adj=1e-04, useFit=F)}
\arguments{
  \item{sem.obj}{A fitted \code{\link{sem}}. object}
  \item{sem.data}{A data frame of variables used to generate the covariance matrix used in fitting the \code{\link{sem}}. object.}
  \item{adj}{The adjustment to each parameter value used to calculate the change in the covariance matrix when running \code{\link{delta_matrix}}..}
  \item{useFit}{Use the observed or fitted covariance matrix for calculation of the maximum liklihood weight matrix.  Defaults to False.}
}

\details{For structural equation models fit to non-normally distributed data, the standard Chi Square goodness of fit statistic (T) can be quite inflated.  This function calculates an adjusted statistic, Tbar, as proposed by Satorra and Bentler (1988) as follows:
\deqn{Tbar = T/c}
\deqn{c = trace(U*G)/df}
	  
	  where df is the model degrees of freedom.  G is the ADF weight matrix as calculated by \code{\link{adf.wmat}}.  U, the residual weight matrix, is defined as:
\deqn{U = W - inv(W * D * (D' * W * D) * D' * W)}

	where W is the normal theory weight matrix (in this case generated by \code{\link{ml.wmat}}.) and D is the model Jacobian matrix as defined by \code{\link{delta_matrix}}..
}

\value{ An object of class \code{\link{adjchisq}}.}

\seealso{
	\code{\link{adf.wmat}}.
	\code{\link{ml.wmat}}.
	\code{\link{delta_matrix}}.
	\code{\link{sem}}.
	\code{\link{adjchisq}}.
}


\references{
Bentler, P.M. and Dudgeon, P.  1996.  Covariance structure analysis: Statistical practice, theory, and directions.  \emph{Annual Review of Psychology}.  47:563-592.

Satorra, A. and Bentler, P. M.  1988.  Scaling corrections for chi-square statistics in covariance structure analysis. American Statistical Association.  \emph{Proceedings of the Business and Economic Statistics Section}.  308-313.

}