\name{sem.additions}
\title{Additional Methods for Structural Equation Modeling}
\alias{sem.additions}

\details{ A number of methods for dealing with structural equation models generated by sem in R}

\description{ This code contains the following:

	Model manipulation routines
	SB Chisq, SB Mean Chi Square, 
	SB Robust Standard Errors and Sandwich Standard Errors
	SB ANOVA



I would like it to also contain the following
	YB Chi Square, BCC, ECVI, and more
	Multigroup Analysis
	Graphical Interface for Model Specification
	Retrospective power analysis
	
	Check out what is here, and if you have code contributions, please contact me at \email{byrnes@msi.ucsb.edu}
}

\seealso{
  \code{\link{sem}}.
}

\author{Jarrett Byrnes \email{byrnes@msi.ucsb.edu}}


