\name{sbDiff}
\alias{sbDiff}
\title{Chi Square Difference Test Using Satorra-Bentler Scaled Chi-Square Values}
\description{Calculated the chi square difference test between two nested models, after the S-B correction has been applied}
\usage{sbDiff(sbobj.A, sbobj.B)}


\arguments{
	\item{sbobj.A}{An object generated by \code{\link{sbchisq}} from an \code{\link{sem}} object}
	\item{sbobj.B}{An object generated by \code{\link{sbchisq}} from an \code{\link{sem}} object.  sbobj.B is either a nested subset of sbobj.A or vice-versa.}
}

\details{Calculates the chi square difference between these two models using the methods of Satorra and Bentler(2001).  We use the output of corrected test statistics as \code{\link{adjchisq}} objects contain all of the metrics necessary for the test.

One could code up something that takes an sem object, an altered RAM model, and a data frame - but really, why bother, as you'll probably want output from all of the intervening steps.
Note: this is deprecated and one should use anova.adjchisq
}


\seealso{
	\code{\link{sem}}
	\code{\link{sbchisq}}
}

\references{
	\link{http://www.statmodel.com/chidiff.shtml}
	
	
	Satorra, A. and P.M. Bentler. 2001 A scaled difference chi-square test statistic for moment structure analysis. \emph{Psychometrika}. 66: 507-514.
}